+++ src/pages/MainPage.tsx
@@ -1,5 +1,6 @@
 import { ChangeEvent, FC, MouseEvent, useEffect, useState } from 'react';
 import TodoItem from '../components/TodoItem';
+import TodoListHeader from '../components/TodoListHeader';
 import { TodoServiceImpl } from '../service/todo';
 
 interface MainPageProps {
@@ -14,35 +15,12 @@ export interface Todo {
 }
 const MainPage: FC<MainPageProps> = ({ todoService }) => {
   const [todoList, setTodoList] = useState<Todo[]>([]);
-  const [title, setTitle] = useState<string>('');
-  const [content, setContent] = useState<string>('');
 
   useEffect(() => {
     todoService.getTodos().then((response) => {
       setTodoList(response.data);
     });
   }, [todoService]);
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const target = e.currentTarget.name;
-    const value = e.currentTarget.value;
-    switch (target) {
-      case 'title':
-        return setTitle(value);
-      case 'content':
-        return setContent(value);
-      default:
-        return;
-    }
-  };
-
-  const handleAdd = (e: MouseEvent<HTMLButtonElement>) => {
-    todoService.createTodo(title, content).then((response) => {
-      setTodoList([...todoList, response.data]);
-      setTitle('');
-      setContent('');
-    });
-  };
 
   const handleDelete = (e: MouseEvent<HTMLButtonElement>) => {
     const id = e.currentTarget.parentElement?.dataset.id!;
@@ -51,25 +29,15 @@ const MainPage: FC<MainPageProps> = ({ todoService }) => {
     });
   };
 
+  function handleCreateTodo(title: string, content: string) {
+    todoService.createTodo(title, content).then((response) => {
+      setTodoList([...todoList, response.data]);
+    });
+  }
+
   return (
     <>
-      <div className="add">
-        <input
-          name="title"
-          type="text"
-          value={title}
-          onChange={handleChange}
-          placeholder="제목"
-        />
-        <input
-          name="content"
-          type="text"
-          value={content}
-          onChange={handleChange}
-          placeholder="내용"
-        />
-        <button onClick={handleAdd}>추가</button>
-      </div>
+      <TodoListHeader onCreateTodo={handleCreateTodo} />
       <div className="lists">
         <ul>
           {todoList.map((todo) => (
